worked out formals issue, compiles, needs testing
I mostly modified MethodInfo and Method Table so that MethodInfo creates
a formals VarTable as it checks for variable clashes in formals.
MethodTable now uses those existing tables to add accessors to both
formals and locals in order. Tere is currently no cap on the number of
formals you can have.